:white_check_mark: Fix: Remove broken test from unplugin (#666)
## Description

Test is broken because the version is hardcoded. This is because we
can't import json with no bundler mode. No idea how to deal with this
other than adding a build step for now just gonna ignore the issue.

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>