:sparkles: Feat(vm): Add precompiles as an option (#664)
## Description

[Norswap requested
precompiles](https://twitter.com/norswap/status/1724944291204665439).
Norswap gets precompiles.

With precompiles one will be able to write arbitrary javascript that
runs when a contract is called. They can even do gross stuff like append
to a dom. Fwiw this will also be how evmts cheat codes work but before
we weren't planning on exposing this to end users.

While implementing this feature I found an annoying bug. If you don't
include an argument in your precompile the entire thing fails to
execute. Opened pr to fix issue here
https://github.com/ethereumjs/ethereumjs-monorepo/pull/3158

The types are currently wrong. It takes an Ethjs address instead of a
HexString address unlike the rest of the evmts pr. This is because
patching the types as I wished was tediously hard as a result of
ethereumjs not exporting specific types I need. Later I will do a pr to
ethereumjs about this. For now we do some typescript magic to infer the
type but because of the type is a union type it's difficult to patch and
I gave up for now.

## Testing

Unit test testing a precompile

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>