✨ feat: Add predeploy option to vm CHA-69 (#777)
## Description

In create Tevm we allow the option to pass in pre-deployed contracts
that are deployed to the _evm. A pre-deploy object is comprised of the
desired deploy address and a TevmContract object/

## Testing

I wrote one test that initializes Tevm with a pre-deploy contract. It
then checks for the contracts existence in the state manager and calls
one of the contracts functions

Your ENS/address: noncents.eth



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Expanded the options for creating an EVM instance with support for
custom predeploys.
- Introduced functionality for defining and creating predeployed
contracts.
- Added new test cases for predeploy interactions and HTTP handler
creation.

- **Enhancements**
- Improved import statement consistency by adding explicit file
extensions.

- **Documentation**
- Updated export statements to reflect changes in file naming
conventions.

- **Continuous Integration**
- Modified GitHub workflow to include documentation generation and
optimized command execution order.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>
Co-authored-by: Will Cory <willcory10@gmail.com>