:wheelchair: Feat: Make @tevm/contracts NodeNext (#792)
## Description

Contracts package not being NodeNext causes issues with interoping with
other packages. It was not nodenext because a type for events from viem
was a deep import and nodenext doesn't allow deep imports. Latest viem
makes it so we can get rid of this deep import and then switch to node
next

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Updated import paths across the application for improved module
resolution and consistency.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>