:sparkles: Feat: Add deployHandler (#1058)
Fixes #1052

TODO
- [ ] Add deploy handler
- [ ] Add deploy handler validator
- [ ] Update `tevm.script` to optionally take constructor args

### TL;DR

This pull request introduces a new `DeployHandler` in the
`actions-types` package. The `DeployHandler` is used to deploy a
contract to TEVM. It also includes the necessary changes in the
`actions` package to handle deployment of ERC20 tokens. Testing
instructions and motivation for the changes are also provided.

### What changed?

The changes include:
- Addition of `DeployHandler` in `actions-types` package
- Implementation of deploying ERC20 token in `actions` package

### How to test?

To test the changes, run the test cases for deploying a contract using
the `DeployHandler` in the `actions-types` package.

### Why make this change?

The motivation behind this change is to enable users to deploy contracts
to TEVM and provide support for deploying ERC20 tokens.

---

## Description

_Concise description of proposed changes_

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced the `createTransaction` function to calculate various fees
and gas limits for transaction creation.
- Introduced a `deployHandler` function for deploying contracts with
data encoding and error handling.
- Added functionality for deploying contracts, interacting with ERC20
contracts, and handling mining operations in a test environment.

- **Documentation**
- Updated the exported types in `abitype.ts` to include new types and
remove obsolete ones.
- Added type aliases `ContractConstructorArgs` and
`EncodeDeployDataParameters` to the API documentation.
- Introduced a new type alias `EncodeDeployDataParameters` with detailed
declarations for encoding deploy data parameters in TypeScript.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: William Cory <williamcory@Williams-MacBook-Pro.local>