Version Packages (next) (#1581)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.

⚠️⚠️⚠️⚠️⚠️⚠️

`main` is currently in **pre mode** so this branch has prereleases
rather than normal releases. If you want to exit prereleases, run
`changeset pre exit` on `main`.

⚠️⚠️⚠️⚠️⚠️⚠️

# Releases
## tevm-run@1.0.0-next.138

### Patch Changes

-   tevm@1.0.0-next.138

## @tevm/cli@1.0.0-next.138

### Patch Changes

-   37d9d8f: Fix \_\_dirname error

## tevm@1.0.0-next.138

### Patch Changes

-   Updated dependencies [37d9d8f]
    -   @tevm/cli@1.0.0-next.138

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>